refactor(practice-client-intakes): Decompose monolithic service into granular, domain-specific services and routes.
