+++ src/modules/practice-client-intakes/handlers.ts
@@ -1,6 +1,6 @@
 import { getLogger } from '@logtape/logtape';
 import type { Context } from 'hono';
-import { practiceClientIntakesRepository } from '@/modules/practice-client-intakes/database/queries/practice-client-intakes.repository';
+
 import {
   getIntakeSettingsRoute,
   createPracticeClientIntakeRoute,
@@ -14,8 +14,13 @@ import {
   updateIntakeTriageStatusRoute,
   convertIntakeRoute,
 } from '@/modules/practice-client-intakes/routes';
-import { practiceClientIntakesService } from '@/modules/practice-client-intakes/services/practice-client-intakes.service';
+import { intakeCheckoutService } from '@/modules/practice-client-intakes/services/intake-checkout.service';
+import { intakeConvertService } from '@/modules/practice-client-intakes/services/intake-convert.service';
+import { intakeCreationService } from '@/modules/practice-client-intakes/services/intake-creation.service';
+import { intakeLifecycleService } from '@/modules/practice-client-intakes/services/intake-lifecycle.service';
+import { intakeQueryService } from '@/modules/practice-client-intakes/services/intake-query.service';
 import type { AppRouteHandler } from '@/shared/types/hono';
+import { getServiceContext } from '@/shared/types/service-context';
 import { response } from '@/shared/utils/responseUtils';
 
 const logger = getLogger(['practice-client-intakes', 'handlers']);
@@ -36,23 +41,25 @@ async function getOptionalSessionUserId(c: Context): Promise<string | null> {
 }
 
 export const getIntakeSettingsHandler: AppRouteHandler<typeof getIntakeSettingsRoute> = async (c) => {
+  const ctx = getServiceContext(c);
   const { slug } = c.req.valid('param');
-  const result = await practiceClientIntakesService.getPracticeClientIntakeSettings(slug);
+
+  const result = await intakeQueryService.getPracticeClientIntakeSettings({ slug }, ctx);
   return response.fromResult(c, result);
 };
 
 export const createPracticeClientIntakeHandler: AppRouteHandler<typeof createPracticeClientIntakeRoute> = async (c) => {
+  const ctx = getServiceContext(c);
   const body = c.req.valid('json');
-  const clientIp = c.req.header('x-forwarded-for') || c.req.header('remote-addr');
-  const userAgent = c.req.header('user-agent');
-  const origin = c.req.header('origin');
 
-  const result = await practiceClientIntakesService.createPracticeClientIntake({
-    ...body,
-    clientIp,
-    userAgent,
-    origin,
-  });
+  const result = await intakeCreationService.createPracticeClientIntake({
+    data: {
+      ...body,
+      clientIp: c.req.header('x-forwarded-for') || c.req.header('remote-addr'),
+      userAgent: c.req.header('user-agent'),
+      origin: c.req.header('origin'),
+    },
+  }, ctx);
 
   return response.fromResult(c, result, 201);
 };
@@ -60,50 +67,59 @@ export const createPracticeClientIntakeHandler: AppRouteHandler<typeof createPra
 export const createPracticeClientIntakeCheckoutSessionHandler: AppRouteHandler<
   typeof createPracticeClientIntakeCheckoutSessionRoute
 > = async (c) => {
+  const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
-  const sessionUserId = await getOptionalSessionUserId(c);
-  const result = await practiceClientIntakesService.createPracticeClientIntakeCheckoutSession({
+
+  const user_id = ctx.userId ?? await getOptionalSessionUserId(c) ?? undefined;
+
+  const result = await intakeCheckoutService.createPracticeClientIntakeCheckoutSession({
     uuid,
-    user_id: sessionUserId ?? undefined,
+    user_id,
     origin: c.req.header('origin'),
-  });
+  }, ctx);
+
   return response.fromResult(c, result, 201);
 };
 
 export const updatePracticeClientIntakeHandler: AppRouteHandler<typeof updatePracticeClientIntakeRoute> = async (c) => {
+  const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
   const body = c.req.valid('json');
-  const result = await practiceClientIntakesService.updatePracticeClientIntake(uuid, body);
+
+  const result = await intakeCreationService.updatePracticeClientIntake({ uuid, data: body }, ctx);
   return response.fromResult(c, result);
 };
 
 export const getPracticeClientIntakeStatusHandler: AppRouteHandler<
-  typeof getPracticeClientIntakeStatusRoute> = async (c) => {
-    const { uuid } = c.req.valid('param');
-    const sessionUserId = await getOptionalSessionUserId(c);
-    const result = await practiceClientIntakesService.getPracticeClientIntakeStatus(uuid, sessionUserId ?? undefined);
-    return response.fromResult(c, result);
-  };
+  typeof getPracticeClientIntakeStatusRoute
+> = async (c) => {
+  const ctx = getServiceContext(c);
+  const { uuid } = c.req.valid('param');
+
+  const requestingUserId = ctx.userId ?? await getOptionalSessionUserId(c) ?? undefined;
+
+  const result = await intakeCheckoutService.getPracticeClientIntakeStatus({ uuid, requestingUserId }, ctx);
+  return response.fromResult(c, result);
+};
 
 export const getPracticeClientIntakePostPayStatusHandler: AppRouteHandler<
-  typeof getPracticeClientIntakePostPayStatusRoute> = async (c) => {
-    const { session_id: sessionId } = c.req.valid('query');
-    const result = await practiceClientIntakesService.getPracticeClientIntakePostPayStatus(sessionId);
-    return response.fromResult(c, result);
-  };
+  typeof getPracticeClientIntakePostPayStatusRoute
+> = async (c) => {
+  const ctx = getServiceContext(c);
+  const { session_id: sessionId } = c.req.valid('query');
+
+  const result = await intakeCheckoutService.getPracticeClientIntakePostPayStatus({ sessionId }, ctx);
+  return response.fromResult(c, result);
+};
 
 export const claimPracticeClientIntakeHandler: AppRouteHandler<typeof claimPracticeClientIntakeRoute> = async (c) => {
-  const { session_id: sessionId } = c.req.valid('json');
-  const sessionUserId = c.get('userId');
+  const ctx = getServiceContext(c);
+  const { session_id } = c.req.valid('json');
 
-  if (!sessionUserId) {
-    return response.unauthorized(c, 'Authentication required to claim intake');
-  }
-
-  const result = await practiceClientIntakesService.claimPracticeClientIntakePayment({
-    session_id: sessionId,
-    user_id: sessionUserId,
-  });
+  const result = await intakeLifecycleService.claimPracticeClientIntakePayment({
+    session_id,
+    user_id: ctx.userId,
+  }, ctx);
 
   return response.fromResult(c, result);
 };
@@ -111,79 +127,52 @@ export const claimPracticeClientIntakeHandler: AppRouteHandler<typeof claimPract
 export const triggerIntakeInvitationHandler: AppRouteHandler<
   typeof triggerIntakeInvitationRoute
 > = async (c) => {
+  const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
-  const sessionUserId = c.get('userId');
 
-  if (!sessionUserId) {
-    return response.unauthorized(c, 'Authentication required to trigger invitation');
-  }
-
-  const result = await practiceClientIntakesService.triggerIntakeInvitation(
+  const result = await intakeLifecycleService.triggerIntakeInvitation({
     uuid,
-    sessionUserId,
-    c.req.raw.headers,
-  );
+    sessionUserId: ctx.userId,
+    origin: c.req.header('origin'),
+  }, ctx);
 
   return response.fromResult(c, result);
 };
 
 export const listIntakesHandler: AppRouteHandler<typeof listIntakesRoute> = async (c) => {
+  const ctx = getServiceContext(c);
   const { practice_id: practiceId } = c.req.valid('param');
   const query = c.req.valid('query');
 
-  const result = await practiceClientIntakesService.listIntakes(practiceId, query);
-
+  const result = await intakeQueryService.listIntakes({ practiceId, query }, ctx);
   return response.fromResult(c, result);
 };
 
 export const convertIntakeHandler: AppRouteHandler<typeof convertIntakeRoute> = async (c) => {
+  const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
   const body = c.req.valid('json');
-  const userId = c.get('userId');
 
-  if (!userId) {
-    return response.unauthorized(c, 'Authentication required to convert intake');
-  }
-
-  // We need organizationId for the service call, so we fetch the intake record first.
-  const intake = await practiceClientIntakesRepository.findById(uuid);
-  if (!intake) {
-    return response.notFound(c, 'Practice client intake not found');
-  }
-
-  const result = await practiceClientIntakesService.convertIntakeToMatter(
+  const result = await intakeConvertService.convertIntakeToMatter({
     uuid,
-    intake.organization_id,
-    userId,
-    body,
-  );
+    organizationId: ctx.organizationId,
+    actorUserId: ctx.userId,
+    data: body,
+  }, ctx);
 
   return response.fromResult(c, result, 201);
 };
 
 export const updateIntakeTriageStatusHandler: AppRouteHandler<typeof updateIntakeTriageStatusRoute> = async (c) => {
+  const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
   const body = c.req.valid('json');
-  const userId = c.get('userId');
 
-  if (!userId) {
-    return response.unauthorized(c, 'Authentication required to update intake triage status');
-  }
+  const result = await intakeQueryService.updateIntakeTriageStatus({
+    uuid,
+    organizationId: ctx.organizationId,
+    data: body,
+  }, ctx);
 
-  const activeOrganizationId = c.get('activeOrganizationId');
-  if (!activeOrganizationId) {
-    return response.forbidden(c, 'No active organization selected');
-  }
-
-  const intake = await practiceClientIntakesRepository.findById(uuid);
-  if (!intake) {
-    return response.notFound(c, 'Practice client intake not found');
-  }
-
-  if (intake.organization_id !== activeOrganizationId) {
-    return response.forbidden(c, 'Access denied');
-  }
-
-  const result = await practiceClientIntakesService.updateIntakeTriageStatus(uuid, activeOrganizationId, body);
   return response.fromResult(c, result);
 };
