+++ src/modules/practice-client-intakes/routes/index.ts
@@ -1,0 +1,3 @@
+export * from './checkout.routes';
+export * from './creation.routes';
+export * from './lifecycle.routes';
