+++ src/modules/practice/handlers.ts
@@ -1,13 +1,14 @@
 import { routes } from '@/modules/practice/routes';
+import { practiceDetailsManagementService } from '@/modules/practice/services/practice-details-management.service';
 import { practiceManagementService } from '@/modules/practice/services/practice-management.service';
 import { practiceQueriesService } from '@/modules/practice/services/practice-queries.service';
-import { AppRouteHandler } from '@/shared/types/hono';
+import type { AppRouteHandler } from '@/shared/types/hono';
 import { getServiceContext } from '@/shared/types/service-context';
 import { response } from '@/shared/utils/responseUtils';
 
 export const listPracticesHandler: AppRouteHandler<typeof routes.listPracticesRoute> = async (c) => {
   const ctx = getServiceContext(c);
-  const result = await practiceQueriesService.listPractices({ requestHeaders: c.req.header() }, ctx);
+  const result = await practiceQueriesService.listPractices(ctx);
   return response.fromResult(c, result);
 };
 
@@ -16,7 +17,6 @@ export const createPracticeHandler: AppRouteHandler<typeof routes.createPractice
   const validatedBody = c.req.valid('json');
   const result = await practiceManagementService.createPractice({
     data: validatedBody,
-    requestHeaders: c.req.header(),
   }, ctx);
   return response.fromResult(c, result, 201);
 };
@@ -25,7 +25,7 @@ export const getPracticeHandler: AppRouteHandler<typeof routes.getPracticeByIdRo
   const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
   const result = await practiceQueriesService.getPracticeById(
-    { organizationId: uuid, requestHeaders: c.req.header() },
+    { organizationId: uuid },
     ctx,
   );
   return response.fromResult(c, result);
@@ -38,7 +38,6 @@ export const updatePracticeHandler: AppRouteHandler<typeof routes.updatePractice
   const result = await practiceManagementService.updatePractice({
     organizationId: uuid,
     data: validatedBody,
-    requestHeaders: c.req.header(),
   }, ctx);
   return response.fromResult(c, result);
 };
@@ -46,8 +45,8 @@ export const updatePracticeHandler: AppRouteHandler<typeof routes.updatePractice
 export const deletePracticeHandler: AppRouteHandler<typeof routes.deletePracticeRoute> = async (c) => {
   const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
-  const result = await practiceManagementService.deletePractice(
-    { organizationId: uuid, requestHeaders: c.req.header() },
+  const result = await practiceDetailsManagementService.deletePractice(
+    { organizationId: uuid },
     ctx,
   );
   return response.fromResult(c, result, 204);
@@ -56,8 +55,8 @@ export const deletePracticeHandler: AppRouteHandler<typeof routes.deletePractice
 export const setActivePracticeHandler: AppRouteHandler<typeof routes.setActivePracticeRoute> = async (c) => {
   const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
-  const result = await practiceManagementService.setActivePractice(
-    { organizationId: uuid, requestHeaders: c.req.header() },
+  const result = await practiceDetailsManagementService.setActivePractice(
+    { organizationId: uuid },
     ctx,
   );
   return response.fromResult(c, result);
@@ -67,7 +66,7 @@ export const getPracticeDetailsHandler: AppRouteHandler<typeof routes.getPractic
   const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
   const result = await practiceQueriesService.getPracticeDetails(
-    { organizationId: uuid, requestHeaders: c.req.header() },
+    { organizationId: uuid },
     ctx,
   );
   return response.fromResult(c, result);
@@ -77,10 +76,9 @@ export const createPracticeDetailsHandler: AppRouteHandler<typeof routes.createP
   const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
   const validatedBody = c.req.valid('json');
-  const result = await practiceManagementService.upsertPracticeDetails({
+  const result = await practiceDetailsManagementService.upsertPracticeDetails({
     organizationId: uuid,
     data: validatedBody,
-    requestHeaders: c.req.header(),
   }, ctx);
   return response.fromResult(c, result, 201);
 };
@@ -89,10 +87,9 @@ export const updatePracticeDetailsHandler: AppRouteHandler<typeof routes.updateP
   const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
   const validatedBody = c.req.valid('json');
-  const result = await practiceManagementService.upsertPracticeDetails({
+  const result = await practiceDetailsManagementService.upsertPracticeDetails({
     organizationId: uuid,
     data: validatedBody,
-    requestHeaders: c.req.header(),
   }, ctx);
   return response.fromResult(c, result);
 };
@@ -100,8 +97,8 @@ export const updatePracticeDetailsHandler: AppRouteHandler<typeof routes.updateP
 export const deletePracticeDetailsHandler: AppRouteHandler<typeof routes.deletePracticeDetailsRoute> = async (c) => {
   const ctx = getServiceContext(c);
   const { uuid } = c.req.valid('param');
-  const result = await practiceManagementService.deletePracticeDetails(
-    { organizationId: uuid, requestHeaders: c.req.header() },
+  const result = await practiceDetailsManagementService.deletePracticeDetails(
+    { organizationId: uuid },
     ctx,
   );
   return response.fromResult(c, result, 204);
