+++ src/modules/practice/services/practice-management.service.ts
@@ -1,6 +1,6 @@
 import { getLogger } from '@logtape/logtape';
 import { omit } from 'es-toolkit/compat';
-import { organizationRepository } from '@/modules/practice/database/queries/organization.repository';
+
 import {
   findPracticeDetailsByOrganization,
 } from '@/modules/practice/database/queries/practice-details.repository';
@@ -10,32 +10,19 @@ import {
   DETAILS_FIELD_KEYS,
   buildPracticeWithDetails,
   upsertDetailsTransaction,
-  buildPracticeDetailsDeletedPayload,
-  findAndDeletePracticeDetails,
 } from '@/modules/practice/services/practice-management.helpers';
-import type {
-  PracticeDetailsResponse,
-} from '@/modules/practice/types/practice-details.types';
 import type {
   CreatePracticeParams,
   UpdatePracticeParams,
-  UpsertPracticeDetailsParams,
 } from '@/modules/practice/types/practice-management.types';
 import type {
   UpdatePracticeRequest,
   PracticeWithDetails,
-  OrganizationRequestParams,
 } from '@/modules/practice/types/practice.types';
 import { practiceValidations } from '@/modules/practice/validations/practice.validation';
 import betterAuthUtils from '@/shared/auth/utils/betterAuthUtils';
 import { db } from '@/shared/database';
-import {
-  PracticeCreated,
-  PracticeUpdated,
-  PracticeDeleted,
-  PracticeDetailsDeleted,
-  PracticeSwitched,
-} from '@/shared/events/definitions';
+import { PracticeCreated, PracticeUpdated } from '@/shared/events/definitions';
 import type { Organization } from '@/shared/types/BetterAuth';
 import type { Result } from '@/shared/types/result';
 import type { ServiceContext } from '@/shared/types/service-context';
@@ -48,14 +35,15 @@ const logger = getLogger(['practice', 'management-service']);
 /**
  * Practice Management Service
  *
- * Handles mutation operations for practices (organizations + details)
+ * Handles create and update of practices (org + details coordination with Better Auth).
+ * For delete, details upsert/delete, and session switching see practice-details-management.service.ts
  */
 export const practiceManagementService = {
   /**
    * Create a new practice
    */
   async createPractice(
-    { data, requestHeaders }: CreatePracticeParams,
+    { data }: CreatePracticeParams,
     ctx: ServiceContext,
   ): Promise<Result<{ practice: PracticeWithDetails }>> {
     const { user } = ctx;
@@ -63,7 +51,7 @@ export const practiceManagementService = {
       const organizationData = omit(data, DETAILS_FIELD_KEYS);
 
       const createResult = await organizationService.createOrganization(
-        { data: organizationData, requestHeaders },
+        { data: organizationData },
         ctx,
       );
 
@@ -87,7 +75,7 @@ export const practiceManagementService = {
           });
         } catch (detailsError) {
           const rollbackResult = await organizationService.deleteOrganization(
-            { organizationId: organization.id, requestHeaders },
+            { organizationId: organization.id },
             ctx,
           );
           if (!rollbackResult.success) {
@@ -125,7 +113,7 @@ export const practiceManagementService = {
    * Update an existing practice (Unified Org + Details)
    */
   async updatePractice(
-    { organizationId, data, requestHeaders }: UpdatePracticeParams,
+    { organizationId, data }: UpdatePracticeParams,
     ctx: ServiceContext,
   ): Promise<Result<{ practice: PracticeWithDetails }>> {
     const { user } = ctx;
@@ -142,14 +130,14 @@ export const practiceManagementService = {
 
       if (hasOrganizationUpdates) {
         const previousOrgResult = await organizationService.getFullOrganization(
-          { organizationId, requestHeaders },
+          { organizationId },
           ctx,
         );
         if (!previousOrgResult.success) return previousOrgResult;
         previousOrganization = previousOrgResult.data;
 
         const updateResult = await organizationService.updateOrganization(
-          { data: { organizationId, data: filteredOrgData }, requestHeaders },
+          { data: { organizationId, data: filteredOrgData } },
           ctx,
         );
 
@@ -157,7 +145,7 @@ export const practiceManagementService = {
         organization = updateResult.data;
       } else {
         const orgResult = await organizationService.getFullOrganization(
-          { organizationId, requestHeaders },
+          { organizationId },
           ctx,
         );
         if (!orgResult.success) return orgResult;
@@ -192,7 +180,6 @@ export const practiceManagementService = {
                     metadata: previousOrganization.metadata ?? undefined,
                   },
                 },
-                requestHeaders,
               },
               ctx,
             );
@@ -232,158 +219,4 @@ export const practiceManagementService = {
     }
   },
 
-  /**
-   * Upsert practice details directly
-   */
-  async upsertPracticeDetails(
-    { organizationId, data, requestHeaders }: UpsertPracticeDetailsParams,
-    ctx: ServiceContext,
-  ): Promise<Result<PracticeDetailsResponse>> {
-    const { user } = ctx;
-    try {
-      const orgResult = await organizationService.getFullOrganization(
-        { organizationId, requestHeaders },
-        ctx,
-      );
-      if (!orgResult.success) return orgResult;
-
-      const organization = await organizationRepository.findById(organizationId);
-
-      const existing = await findPracticeDetailsByOrganization(organizationId);
-
-      const result = await db.transaction(async (tx) => {
-        return upsertDetailsTransaction(tx, ctx, {
-          organizationId,
-          userId: user.id,
-          data,
-          existingAddressId: existing?.address_id,
-          isCreate: !existing,
-        });
-      });
-
-      const responseData: PracticeDetailsResponse = {
-        ...result.details,
-        organization_id: organizationId,
-        address: result.addressResult
-          ? {
-            line1: result.addressResult.line1 ?? undefined,
-            line2: result.addressResult.line2 ?? undefined,
-            city: result.addressResult.city ?? undefined,
-            state: result.addressResult.state ?? undefined,
-            postal_code: result.addressResult.postal_code ?? undefined,
-            country: result.addressResult.country ?? undefined,
-          }
-          : null,
-        services: result.syncedServices.map((s) => ({ id: s.id, name: s.name, key: s.key })),
-        name: orgResult.data.name,
-        logo: orgResult.data.logo ?? null,
-        payment_link_enabled: organization?.paymentLinkEnabled ?? false,
-        payment_link_prefill_amount: organization?.paymentLinkPrefillAmount ?? 0,
-      };
-
-      return ok<PracticeDetailsResponse>(responseData);
-    } catch (error) {
-      logger.error(
-        'Failed to upsert practice details for {organizationId}: {error}',
-        { organizationId, error },
-      );
-      return internalError<PracticeDetailsResponse>('Failed to save practice details');
-    }
-  },
-
-  /**
-   * Delete a practice and its details
-   */
-  async deletePractice(
-    { organizationId, requestHeaders }: OrganizationRequestParams,
-    ctx: ServiceContext,
-  ): Promise<Result<{ success: boolean }>> {
-    const { user } = ctx;
-    try {
-      const orgResult = await organizationService.getFullOrganization(
-        { organizationId, requestHeaders },
-        ctx,
-      );
-      if (!orgResult.success) return orgResult;
-
-      const existing = await findPracticeDetailsByOrganization(organizationId);
-
-      const deleteResult = await organizationService.deleteOrganization(
-        { organizationId, requestHeaders },
-        ctx,
-      );
-      if (!deleteResult.success) return deleteResult;
-
-      if (existing) {
-        await ctx.emit(PracticeDetailsDeleted, buildPracticeDetailsDeletedPayload(existing));
-      }
-
-      await ctx.emit(PracticeDeleted, {
-        organization_id: organizationId,
-        had_practice_details: !!existing,
-        practice_details_id: existing?.id,
-        user_email: user.email,
-      });
-
-      return ok<{ success: boolean }>({ success: true });
-    } catch (error) {
-      logger.error('Failed to delete practice {organizationId}: {error}', { organizationId, error });
-      return internalError<{ success: boolean }>('Failed to delete practice');
-    }
-  },
-
-  /**
-   * Delete practice details directly
-   */
-  async deletePracticeDetails(
-    { organizationId, requestHeaders }: OrganizationRequestParams,
-    ctx: ServiceContext,
-  ): Promise<Result<{ success: boolean }>> {
-    try {
-      const orgResult = await organizationService.getFullOrganization(
-        { organizationId, requestHeaders },
-        ctx,
-      );
-      if (!orgResult.success) return orgResult;
-
-      await findAndDeletePracticeDetails(ctx, organizationId);
-
-      return ok<{ success: boolean }>({ success: true });
-    } catch (error) {
-      logger.error(
-        'Failed to delete practice details for {organizationId}: {error}',
-        { organizationId, error },
-      );
-      return internalError<{ success: boolean }>('Failed to delete practice details');
-    }
-  },
-
-  /**
-   * Set active practice for the session
-   */
-  async setActivePractice(
-    { organizationId, requestHeaders }: OrganizationRequestParams,
-    ctx: ServiceContext,
-  ): Promise<Result<{ success: boolean }>> {
-    const { user } = ctx;
-    try {
-      const activeResult = await organizationService.setActiveOrganization(
-        { organizationId, requestHeaders },
-        ctx,
-      );
-      if (!activeResult.success) return activeResult;
-
-      await ctx.emit(PracticeSwitched, {
-        user_id: user.id,
-        to_organization_id: organizationId,
-        user_email: user.email,
-        switched_to_organization: organizationId,
-      });
-
-      return ok<{ success: boolean }>({ success: true });
-    } catch (error) {
-      logger.error('Failed to set active practice {organizationId}: {error}', { organizationId, error });
-      return internalError<{ success: boolean }>('Failed to set active practice');
-    }
-  },
 };
